remove appaloon word

+++ controller/controller-menu-page.php
@@ -3,8 +3,8 @@
 
 namespace wp_gdpr\controller;
 
-use wp_gdpr\lib\Appsaloon_Customtables;
-use wp_gdpr\lib\Appsaloon_Table_Builder;
+use wp_gdpr\lib\Gdpr_Customtables;
+use wp_gdpr\lib\Gdpr_Table_Builder;
 use wp_gdpr\lib\Gdpr_Container;
 
 class Controller_Menu_Page {
@@ -25,13 +25,14 @@ class Controller_Menu_Page {
 	 * delete all comments selected in admin menu in form
 	 */
 	public function delete_comments() {
-		if ( 'POST' == $_SERVER['REQUEST_METHOD'] && isset( $_REQUEST['gdpr_delete_comments'] ) && is_array( $_REQUEST['gdpr_requests'] ) ) {
+		if ( 'POST' == $_SERVER['REQUEST_METHOD'] && isset( $_REQUEST['gdpr_delete_comments'] ) && isset( $_REQUEST['gdpr_requests'] ) && is_array( $_REQUEST['gdpr_requests'] ) ) {
 			foreach ( $_REQUEST['gdpr_requests'] as $single_request_id ) {
 				$single_request_id  = sanitize_text_field( $single_request_id );
 				$comments_to_delete = $this->find_delete_request_by_id( $single_request_id );
 
 				$this->unserialize_array_and_delete_comments( $comments_to_delete['comments'] );
 				$this->delete_delete_request( $single_request_id );
+				$this->set_notice(__( 'Comments deleted', 'wp_gdpr' ));
 			}
 		}
 	}
@@ -46,7 +47,7 @@ class Controller_Menu_Page {
 	public function find_delete_request_by_id( $id ) {
 		global $wpdb;
 
-		$table_name = $wpdb->prefix . Appsaloon_Customtables::DELETE_REQUESTS_TABLE_NAME;
+		$table_name = $wpdb->prefix . Gdpr_Customtables::DELETE_REQUESTS_TABLE_NAME;
 
 		$query  = "SELECT * FROM $table_name WHERE ID='$id'";
 		$result = $wpdb->get_results( $query, ARRAY_A );
@@ -77,7 +78,7 @@ class Controller_Menu_Page {
 	 */
 	public function delete_delete_request( $request_id ) {
 		global $wpdb;
-		$table_name = $wpdb->prefix . Appsaloon_Customtables::DELETE_REQUESTS_TABLE_NAME;
+		$table_name = $wpdb->prefix . Gdpr_Customtables::DELETE_REQUESTS_TABLE_NAME;
 		$args       = array( 'ID' => $request_id );
 		$wpdb->delete( $table_name, $args );
 	}
@@ -99,7 +100,7 @@ class Controller_Menu_Page {
 		//add checkbox input in every element with e-mail address
 		$requesting_users = array_map( array( $this, 'map_checkboxes_send_email' ), $requesting_users );
 		//show table object
-		$table = new Appsaloon_Table_Builder(
+		$table = new Gdpr_Table_Builder(
 			array(
 				__( 'id', 'wp_gdpr' ),
 				__( 'e-mail', 'wp_gdpr' ),
@@ -121,7 +122,7 @@ class Controller_Menu_Page {
 	public function get_requests_from_gdpr_table() {
 		global $wpdb;
 
-		$table_name = $wpdb->prefix . Appsaloon_Customtables::REQUESTS_TABLE_NAME;
+		$table_name = $wpdb->prefix . Gdpr_Customtables::REQUESTS_TABLE_NAME;
 
 		$query = "SELECT * FROM $table_name";
 
@@ -192,7 +193,7 @@ class Controller_Menu_Page {
 	public function print_inputs_with_emails() {
 		global $wpdb;
 
-		$table_name = $wpdb->prefix . Appsaloon_Customtables::REQUESTS_TABLE_NAME;
+		$table_name = $wpdb->prefix . Gdpr_Customtables::REQUESTS_TABLE_NAME;
 
 		$query = "SELECT * FROM $table_name";
 
@@ -229,7 +230,7 @@ class Controller_Menu_Page {
 
 				$content = $this->get_email_content( $request[0]['email'],  $request[0]['timestamp'] );
 
-				$this->set_notice();
+				$this->set_notice(__( 'E-mail send', 'wp_gdpr' ));
 
 
 				wp_mail( $to, $subject, $content, array() );
@@ -281,12 +282,12 @@ class Controller_Menu_Page {
 		return home_url() . '/' . base64_encode( 'gdpr#' . $email . '#' . base64_encode( $timestamp ) );
 	}
 
-	public function set_notice() {
+	public function set_notice( $message ) {
 		/**
 		 * set notice
 		 */
-		$notice = Gdpr_Container::make( 'wp_gdpr\lib\Appsaloon_Notice' );
-		$notice->set_message( __( 'E-mail send', 'wp_gdpr' ) );
+		$notice = Gdpr_Container::make( 'wp_gdpr\lib\Gdpr_Notice' );
+		$notice->set_message($message  );
 		$notice->register_notice();
 	}
 
@@ -304,7 +305,7 @@ class Controller_Menu_Page {
 	public function build_table_with_delete_requests() {
 
 		global $wpdb;
-		$table_name = $wpdb->prefix . Appsaloon_Customtables::DELETE_REQUESTS_TABLE_NAME;
+		$table_name = $wpdb->prefix . Gdpr_Customtables::DELETE_REQUESTS_TABLE_NAME;
 
 		$query = "SELECT * FROM $table_name";
 
@@ -312,7 +313,7 @@ class Controller_Menu_Page {
 		$requests = array_map( array( $this, 'add_delete_checkbox' ), $requests );
 		$requests = array_map( array( $this, 'reduce_comments_to_string' ), $requests );
 
-		$table = new Appsaloon_Table_Builder(
+		$table = new Gdpr_Table_Builder(
 			array(
 				__( 'id', 'wp_gdrp' ),
 				__( 'e-mail', 'wp_gdrp' ),
@@ -372,7 +373,7 @@ class Controller_Menu_Page {
 
 		$plugins = $this->filter_plugins( $plugins );
 
-		$table = new Appsaloon_Table_Builder(
+		$table = new Gdpr_Table_Builder(
 			array( __( 'plugin name', 'wp_gdpr' ) ),
 			$plugins
 			, array() );
